backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
orjson==3.10.7
python-dotenv==1.0.1
SQLAlchemy==2.0.32
aiosqlite==0.20.0
# PostgreSQL (optional): pip install -r requirements-postgres.txt — use Python 3.11/3.12 for asyncpg
# asyncpg==0.29.0
# psycopg[binary,pool]>=3.2.2
# databases[postgresql]==0.9.0
redis==5.0.8
websockets==12.0
# protobuf version chosen by ortools
ortools>=9.12.4544
torch>=2.6.0
# ray[rllib]  # No Python 3.13 wheel; uncomment on Python 3.11/3.12 for RL features
networkx==3.3
numpy>=1.26.4
pandas>=2.2.2
tenacity==8.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

